added another file for the second class
